Add STM32F04x option byte dumps for use with dfu-util
factory_default.options.bin contains the default option bytes for
the STM32F042x - no read/write protection and BOOT0 can be used to
enable the DFU bootloader.

no_boot_pin.options.bin has the BOOT_SEL bit cleared so that BOOT0
is ignored. This allows the bootloader pin to be used with inputs
that could be high when the board is reset (for example CAN_RX).

The current option bytes can be read out using dfu-util:

    dfu-util -a 1 -s 0x1FFFF800 -U dest-filename-goes-here

Writing the option bytes can be done with the -D switch:

    dfu-util -a 1 -s 0x1FFFF800 -D src-filename-goes-here