composite_usb_conf: add bulk to packet memory calculations
Packet memory is a finite resource. This is memory where packet data
is buffered as it goes to the host, or as packets are received from
the host.

Previously, bulk endpoints were not accounted for. This didn't cause
any observable issues because there is enough memory. Add the check
for bulk data in case this is ported to a device that has a packet
memory area that's too small.

Signed-off-by: Sean Cross <sean@xobs.io>