bulk: support pipelining bulk IN transfers
Support having more than one bulk response in flight at a time. A host can
make multiple requests without reading the response. In this case, store it
in the ring buffer just like we do with HID.

As part of this, it was discovered that the ZLP handling was incorrect, and
the feature was removed entirely.

Signed-off-by: Sean Cross <sean@xobs.io>