tinydyne: add new target
Tinydyne is a simple all-in-one development board. It is connected to an
STM32F411 via UART, SWD, and JTAG. As a result, it can query the target via
either JTAG or SWD.

Both communication mechanisms are verified to work.

SWO is connected to PB7, which is also USART1_RX. SWO can be added in a
future version.

Signed-off-by: Sean Cross <sean@xobs.io>